implement edit HEX color manually in ColorPicker component
